Adding Mastodon support
A new command will scout the feed's homepages for links to Mastodon
profiles (marked with rel=me), verifies that the link indeed points to a
Mastodon profile and also checks that the profile links back to the
homepage. These verified Mastodon accounts are added to the database.

Another new command will post a random post to a configured Mastodon
account. If we have a mastodon profile for the item's feed, it will be
mentioned in the toot.